Update index.html
Tweaked paddle collision algorithm in an effort to reduce instances of ball sticking inside of paddle. Added an if / else statement on the button listener to check if game.active is false, and to change game.active to true if it's the first time clicking the button. Changed ball tracking for paddleTwo to keep paddleTwo from moving outside of the frame.